# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SuperlativeBridge API",
    "version": "1.2.0",
    "description": "REST API backing the SuperlativeBridge platform. Runs on AWS Lambda + API Gateway in production and as a single Go HTTP server locally (docker compose). Data lives in Postgres.\n\nSeeded demo accounts (password: `password123`): admin@example.com, maria@example.com (GIG worker), aisha@example.com (employer), sarah.mentor@example.com (mentor)."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same origin (local dev proxy)"
    },
    {
      "url": "http://localhost:8081",
      "description": "Local Go API"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
  "paths": {
...
    "/api/users/{id}/enrollments": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get enrollments for the given id",
        "operationId": "get_api_users_id_enrollments",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "403": {
            "description": "Forbidden for this role"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/users/{id}/certifications": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get certifications for the given id",
        "operationId": "get_api_users_id_certifications",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "403": {
            "description": "Forbidden for this role"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/users/{id}/notifications": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get notifications for the given id",
        "operationId": "get_api_users_id_notifications",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "403": {
            "description": "Forbidden for this role"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/users/{id}/dashboard": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get dashboard for the given id",
        "operationId": "get_api_users_id_dashboard",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "403": {
            "description": "Forbidden for this role"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/users/{id}/saved-courses": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get saved-courses for the given id",
        "operationId": "get_api_users_id_saved_courses",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "403": {
            "description": "Forbidden for this role"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/users/{id}/mentor-bookings": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get mentor-bookings for the given id",
        "operationId": "get_api_users_id_mentor_bookings",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "403": {
            "description": "Forbidden for this role"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/enrollments": {
      "post": {
        "tags": [
          "enrollments"
        ],
        "summary": "Create enrollment",
        "operationId": "post_api_enrollments",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "403": {
            "description": "Forbidden for this role"
          },
          "404": {
            "description": "Not found"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/saved-courses": {
      "post": {
        "tags": [
          "savedcourses"
        ],
        "summary": "Create saved-course",
        "operationId": "post_api_saved_courses",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "403": {
            "description": "Forbidden for this role"
          },
          "404": {
            "description": "Not found"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/mentor-bookings": {
      "post": {
        "tags": [
          "mentorbookings"
        ],
        "summary": "Create mentor-booking",
        "operationId": "post_api_mentor_bookings",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "403": {
            "description": "Forbidden for this role"
          },
          "404": {
            "description": "Not found"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/users/{id}/analytics": {
      "get": {
        "tags": ["users"],
        "summary": "Detailed analytics for one user",
        "operationId": "get_api_users_id_analytics",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "User metrics, enrollments, bookings, certifications",
            "content": { "application/json": { "schema": { "type": "object" } } }
          }
        }
      }
    },
...
  }
}